An attempt to deal with memory leaks - this is a patch I found in ng-grid's issues:
https://github.com/angular-ui/ng-grid/issues/1164

It seems to be related to our problems with references to the ng-grid's scope still remaining.

When this patch is applied, we see some good results - heap snapshots taken in a page with an ng-grid, before and after navigating from/to it, are practically the same.
That was not the case before this patch.
So we assume memory leaks that were, are solved with this patch.