A fix to bugs related to selections and sorting. This is a crucial change, so a long explanation will follow :)
When the data of ng-grid is changed, there is a concern to maintain the rows that were already built for the data, build new rows if there's new items in the data, and maintain the state of the rows (mainly - their selection state), after sorting the new data.
There seemed to be a problematic implementation of this.
ng-grid attempted to hold a mapping between the items in the data, and where their rows are located after the sort.
But this mapping assumed that the new data that has just arrived correlates with the data that was previously in the ng-grid.
That's a problematic assumption...
What we do now, is to try to re-organize the rows, with or without sorting, and match them with the new data based on the primary key set for the items in the data.