Fix selection going haywire with big selection
When a lot of rows were selected, where "a lot" means more than the row
virtualization limit, deselecting and reselecting would get screwed up.
Easiest way to reproduce would be to have a table with ~50-100 rows,
select all, and then start trying to deselect rows on random scrolling
positions. It would usually not work.

This was caused due to a bug where ng-grid's row cloning mechanism would
get out of sync. Row X has a clone of it, Y. Later we change Y to be a
clone of row Z, but no one tells row X it has no clone now. So when we
try to change the selection of row X, it messes with Y too, which gets
things out of sync on screen.

My workaround was to override the setter for the `orig` property on
clones and make sure that if it's changed we clean up our trail.